[課題F-3]Webアプリ
投稿時間も表示できるように設計しました

+++ bbs.py
@@ -1,14 +1,16 @@
 from flask import Flask, render_template, request
 import csv
-
+import datetime
+import os
 
 app = Flask(__name__)
 
 
 def main():
-    with open('bbs.csv', 'w') as f:
-        writer = csv.writer(f)
-        writer.writerow(["username", "message"])
+    if(os.path.isfile('bbs.csv') == False):
+        with open('bbs.csv', 'w') as f:
+            writer = csv.writer(f)
+            writer.writerow(["username", "message", "time"])
 
 
 @app.route('/', methods=['POST', "GET"])
@@ -22,6 +24,8 @@ def bbs():
             bbs_dic = {}
             bbs_dic["name"] = row["username"]
             bbs_dic["message"] = row["message"]
+            bbs_dic["time"] = row["time"]
+
             bbs_list.append(bbs_dic)
 
     if request.method == "GET":
@@ -33,13 +37,22 @@ def bbs():
         if name == "":
             name ="名無しさん"
         message = request.form['umessage']
+        now = datetime.datetime.now()
+        month = '{0:%m}月'.format(now)
+        day = '{0:%m}日'.format(now)
+        hour = '{0:%H}時'.format(now)
+        minute = '{0:%M}分'.format(now)
+        second = '{0:%M}秒'.format(now)
+        time = f"{month}{day}{hour}{minute}{second}"
+
         bbs_dic = {}
         bbs_dic["name"] = name
         bbs_dic["message"] = message
+        bbs_dic["time"] = time
         bbs_list.append(bbs_dic)
         with open('bbs.csv', 'a') as f:
             writer = csv.writer(f)
-            writer.writerow([name, message])
+            writer.writerow([name, message, time])
 
     return render_template('bbs.html', username=name, umessage=message, message=bbs_list)
